Prepend project dir to media dir
Before, when trying to pull staging media or images with `fab pull-staging-media`
or `fab pull-staging-images`, the command would fail because of missing
permissions to write to the `media` directory.

This made sense, because the fabfile was setting the `LOCAL_MEDIA_FOLDER = media`.
This would mean that the command is trying to write to the system directory
`/media`.

This would not work, because the app user `tbx` is not allowed to override this
directory. Nor should that directory be used, as this is for media mounts.

The directory that should be used is `/app/media`. This is clear from the
`docker-compose` file where the `/app/media` directory is mounted as a volume,
 which allows the downloaded images to be accesses by the host and other containers.

This also means, that even if we would change the permissions of the `/media`
directory,  so that we can write the downloaded files into it, the files would
not be accessible in other containers.