Prefix database_dumps folder in container
This is to fix a similar issue as with the media directory. Becuase the
`database_dumps` directory was created in the root, it was belonging to the
root user. This meant that the `tbx` user was not able to write the database
dumps in to the folder.

The folder is now prefixed with `/app`, which is owned by the `tbx` user.

Therefore it is now possble for the `tbx` user in the util container to write
the database dump into that directory.

To also be able to restore the database dump from the correct location, the
mounted volume on the database container needed to be adjusted.