Chore: Upgrade Django 3.2 and Wagtail 2.15 (#369)
* Upgrade Django to 3.0

* Replace `ugettext` with `gettext`

django.utils.translation.ugettext(), ugettext_lazy(), ugettext_noop(),
ungettext(), and ungettext_lazy() are deprecated in favor of the
functions that they’re aliases for: django.utils.translation.gettext(),
gettext_lazy(), gettext_noop(), ngettext(), and ngettext_lazy().

See also:
https://docs.djangoproject.com/en/3.2/releases/3.0/#id3

* Remove unused and deprecated template tag

The `staticfiles` template tag has been removed in Django 3.0 and is
deprecated since 2.1. The replacement would be `static`.

But since no static template tag is actually used in the template, the
loading can be removed.

* Upgrade Django to 3.1

* Replace django.conf.urls.url with django.urls.re_path

As of Django 3.1, django.conf.urls.url() alias of django.urls.re_path() is deprecated.

See also:
https://docs.djangoproject.com/en/3.2/releases/3.1/#id2

* Update django-pattern-library

This is to fix the deprecation warnings:
```
/usr/local/lib/python3.8/site-packages/pattern_library/urls.py:10: RemovedInDjango40Warning: django.conf.urls.url() is deprecated in favor of django.urls.re_path()
```

Because django-pattern-library and wagtail-markdown depend on markdown,
wagtail-markdown had to be updated.

Then there is an additional issue with markdown and poetry, which
required the exact version to be pinned in the pyproject.toml

* Remove mount of non-existent file

This was causing me issues with isort becase isort will use `isort.cfg`
for configuration. At some point there was a `setup.cfg` directory
created which caused isort to fail.

I hope this will prevent that strange directory from being created
again.

* Fix import order

* Remove markdown pinned version

This is not necessart anymore since a fix was released for `markdown`
that prevents the issue that required the pin.

* Upgrade Django to 3.2

* Define default auto field type

As of Django 3.2, the type of the primary key auto field needs to be defined
either gloablly or in each AppConfig.

See also: https://docs.djangoproject.com/en/3.2/releases/3.2/#customizing-type-of-auto-created-primary-keys

* Remove default_app_config

As of Django 3.2, it is not necessary anymore to declare the default app
config, because the app config is now discovered automatically.

See also:
https://docs.djangoproject.com/en/3.2/releases/3.2/#automatic-appconfig-discovery

* Upgrade django-taggit to fix deprecation warning

This upgrade fixes the `default_app_config` deprecation.

See also:
https://docs.djangoproject.com/en/3.2/releases/3.2/#automatic-appconfig-discovery
https://github.com/jazzband/django-taggit/commit/5764d4dda1ce682d7bcb8a376bd26b93c3574836#diff-20d06ac849f63fd67c6d1536546777d5ac6e3bd1b67118dc6ea5b9acd096fca2

* Upgrade scout-apm to fix default_app_config deprecation

This upgrade fixes the Django 3.2 deprecation warning:

```
/usr/local/lib/python3.8/site-packages/django/apps/registry.py:91: RemovedInDjango41Warning: 'scout_apm.django' defines default_app_config = 'scout_apm.django.apps.ScoutApmDjangoConfig'. Django now detects this configuration automatically. You can remove default_app_config.
```

See also:

* https://docs.djangoproject.com/en/3.2/releases/3.2/#automatic-appconfig-discovery
* https://github.com/scoutapp/scout_apm_python/blob/master/CHANGELOG.md#2190-2021-05-04

* Upgrade Wagtail to 2.14

* Fix modelcluster imports

Before, the parental key fields had been imported from Wagtail. These
fields are only imported by Wagtail and that is apparently not possible
anymore.

The import has been updated to use the original import from
django-modelcluster.

* Upgrade Wagtail to 2.15

* Use new search backend

As of Wagtail 2.15, the previously used postgres search backend module
is deprecated in favour of a unified search backend interface.

See also:
https://docs.wagtail.io/en/stable/releases/2.15.html#database-search-backends-replaced